+++ src/components/update-info.ts
@@ -20,16 +20,16 @@ class UpdateInfo extends ModalWindow {
             visibilityProperty: 'updateInfoVisible',
         });
     }
-    static version = '1.5';
+    static version = '1.5.1';
     view() {
         const content = html`
             <div class="update-info">
                 <div>
-                    <h1>Kahuna has been updated to version ${build.version} 🎉</h1>
+                    <h1>Kahuna has been updated to version ${build.version} 🌞</h1>
                     ${paragraph(this.paragraphs[0], this.links())}
                     ${paragraph(this.paragraphs[1], this.links())}
                     ${paragraph(this.paragraphs[2], this.links())}
-                    <p>Have fun and happy building!</p>
+                    <p>Have a good day and happy building!</p>
                     <div class="center">
                         ${button({ content: 'close', '@click': super.close.bind(this) })}
                     </div>
@@ -39,9 +39,12 @@ class UpdateInfo extends ModalWindow {
         return super.node(content);
     }
     readonly paragraphs = [
-        `The highlights of this release include a brand-new database [schema editor],
-        as well as handy tools for [copying databases] and [copying tables]. The full
-        list of changes and bug fixes can be found in the [CHANGELOG].`,
+        `I felt like it was time for a maintenance release and so the bug fixes and
+        smaller improvements from the last 2 months are collected in this version.
+        As an immediately visible change, strings and numbers are now highlighted
+        in color when displayed. The colors used can be [configured] according to
+        personal preferences. The full list of changes and bug fixes can be found
+        in the [CHANGELOG].`,
         `I'd love to hear your feedback! If you enjoy using Kahuna, please consider
         rating on the [Chrome Web Store] or the [Firefox Add-ons] site. Or show your
         support with a ⭐ on [GitHub]!`,
@@ -52,16 +55,8 @@ class UpdateInfo extends ModalWindow {
     links(): Map<string, string> {
         return new Map([
             [
-                'schema editor',
-                'https://hummingme.github.io/kahuna-docs/database/#edit-schema',
-            ],
-            [
-                'copying databases',
-                'https://hummingme.github.io/kahuna-docs/database/#copy-database',
-            ],
-            [
-                'copying tables',
-                'https://hummingme.github.io/kahuna-docs/datatable/#copy-table',
+                'configured',
+                'https://hummingme.github.io/kahuna-docs/configuration/#application',
             ],
             ['changelog', 'https://github.com/hummingme/kahuna/blob/main/CHANGELOG.md'],
             [